backendDjango-main/pyproject.toml
[tool.poetry]
name = "backenddjango-main"
version = "0.1.0"
description = ""
authors = ["CarlosEduardoCardozo <cardozogumsc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
